requirements.txt
numpy
pandas
yfinance
duckduckgo-search
torch
scikit-learn
python-dotenv
google-genai
streamlit
groq